Code used to solve day 20, part 2
By inspection, I found that rx is fed by a conjunction module kc. I made a hypothesis that each input to kc because true on a relatively short loop, so finding the LCM of those loops would give a final solution. I printed whenever kc got a high pulse, confirmed the hypothesis, and calculated the lcm in python, which gave the result.